backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
qrcode[pil]==7.4.2
gunicorn==21.2.0
python-dotenv==1.0.0